Move <ProseMirror> component to its own module
